% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first.R
\name{convex_setup}
\alias{convex_setup}
\title{Doing the setup for the package convexjlr}
\usage{
convex_setup(backend = c("XRJulia", "JuliaCall"))
}
\arguments{
\item{backend}{whether to use XRJulia or JuliaCall as backend}
}
\description{
This function does the setup for the package convexjlr.
Firstly it will try to establish the connect to Julia via the XRJulia interface,
Secondly it will check for the Julia packages Convex and SCS,
if the packages are not found, it tries to install them into Julia.
Finally, it will try to load the Julia packages and do the neccessary initial setup.
}
\examples{
\dontrun{
convex_setup()
}
}
