% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{cvx_optim}
\alias{cvx_optim}
\title{Solve optimization problem}
\usage{
cvx_optim(p, max_iters = 20000)
}
\arguments{
\item{p}{optimization problem to be solved.}

\item{max_iters}{the maximal iteration times of the convex problem solver,
users may need to increase the maximal iteration times
if the optimal can't be achieved with the default 20000.}
}
\value{
status of optimized problem.
}
\description{
\code{cvx_optim} solves optimization problem using Convex.jl.
}
\examples{
\dontrun{
    convex_setup()
    x <- Variable()
    b <- 1
    p <- minimize(sum((x - b) ^ 2))
    cvx_optim(p)
}
}
