% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{dot}
\alias{dot}
\title{Inner product}
\usage{
dot(x, y)
}
\arguments{
\item{x}{input vector, one input vector need to be constant.}

\item{y}{input vector, one input vector need to be constant.}
}
\description{
Inner product of two input vectors.
}
