% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{dotsort}
\alias{dotsort}
\title{Inner product of two vectors after sorted}
\usage{
dotsort(x, y)
}
\arguments{
\item{x}{input vector, one input vector needs to be constant.}

\item{y}{input vector, one input vector needs to be constant.}
}
\description{
Inner product of two input vectors after sorted.
}
