% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{Expr}
\alias{Expr}
\title{Create expressions to be used for optimization problem creation}
\usage{
Expr(x)
}
\arguments{
\item{x}{expression to be created.}
}
\description{
\code{Expr} create expressions, which can be used later for problem creation.
}
\examples{
\dontrun{
    convex_setup()
    x <- Variable(2)
    x1 <- Expr(x + 1)
}
}
