% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{matrixfrac}
\alias{matrixfrac}
\title{x^T P^{-1} x}
\usage{
matrixfrac(x, P)
}
\arguments{
\item{x}{input vector.}

\item{P}{input matrix, needs to be positive semidefinite.}
}
\description{
x^T P^{-1} x.
}
