% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{quadform}
\alias{quadform}
\title{x^T P x}
\usage{
quadform(x, P)
}
\arguments{
\item{x}{input vector, either x or P must be constant.}

\item{P}{input matrix, either x or P must be constant, P needs to be semidefinte if x is not constant.}
}
\description{
x^T P x.
}
