% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{sumlargest}
\alias{sumlargest}
\title{Sum of the largest elements}
\usage{
sumlargest(x, k)
}
\arguments{
\item{x}{input vector.}

\item{k}{a positive integer.}
}
\description{
Sum of k largest elements of input vector x.
}
