% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{sumsmallest}
\alias{sumsmallest}
\title{Sum of the smallest elements}
\usage{
sumsmallest(x, k)
}
\arguments{
\item{x}{input vector.}

\item{k}{a positive integer.}
}
\description{
Sum of k smallest elements of input vector x.
}
