% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first.R
\name{variable_creating}
\alias{variable_creating}
\alias{Variable}
\alias{Semidefinite}
\title{Create variable for optimization problem}
\usage{
Variable(size = 1, sign = c("None", "Positive", "Negative"))

Semidefinite(size = 1, sign = c("None", "Positive", "Negative"))
}
\arguments{
\item{size}{variable size.}

\item{sign}{whether variable is element-wise positive, element-wise negative
or neither.}
}
\description{
Create variable (vector, matrix, semidefinite matrix and etc.)
for optimization problem.
}
\examples{
x <- Variable(4)
X <- Variable(c(4, 4), sign = "Positive")
S <- Semidefinite(4)
}
