% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{vecdot}
\alias{vecdot}
\title{Inner product of vector representation of two matrices}
\usage{
vecdot(x, y)
}
\arguments{
\item{x}{input matrix, one input matrices need to be constant.}

\item{y}{input matrix, one input matrices need to be constant.}
}
\description{
Inner product of vector representation of two input matrices.
}
