% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{vecnorm}
\alias{vecnorm}
\title{p-norm of vector representation of x}
\usage{
vecnorm(x, p = 2)
}
\arguments{
\item{x}{input matrix.}

\item{p}{a number greater than 1.}
}
\description{
p-norm of vector representation of x, which is deprecated,
use \code{norm(vec(x), p=2)} instead.
}
