% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first.R
\name{convex_setup}
\alias{convex_setup}
\title{Doing the setup for the package convexjlr}
\usage{
convex_setup(backend = c("XRJulia", "JuliaCall"), JULIA_HOME = NULL)
}
\arguments{
\item{backend}{whether to use XRJulia or JuliaCall as backend}

\item{JULIA_HOME}{the path to julia binary,
if not set, convexjlr will try to use the julia in path.}
}
\description{
This function does the setup for the package convexjlr.
Firstly it will try to establish the connect to Julia via the XRJulia interface,
or try to embed julia in R through JuliaCall.
Secondly it will check for the Julia packages Convex and SCS,
if the packages are not found, it tries to install them into Julia.
Finally, it will try to load the Julia packages and do the necessary initial setup.
}
\examples{
\dontrun{
convex_setup()
}
}
